IMPALA-4530: Decrease back-pressure in sorter
Work in progress - POC implementation
Introduces a new in-memory merge level between the quicksort and the external merge phase.
Produces many smaller in-memory runs, sorts those with quicksort, then merges them
in memory, before spilling or GetNext().
Works with RUN_SIZE development query option that determines the number of pages in a run.
The exact run size is to be determined later.
LIMIT_SORT_RUN_BYTES_INITIALLY QO triggers spilling at SORT_RUN_BYTES_LIMIT from the
beginning to help measuring the effect on back-pressure. This flag is only for development.
Needs additional page(s) because of the in-memory merger, thus MinReservation is increased.

Change-Id: I58c0ae112e279b93426752895ded7b1a3791865c